+++ lib/src/stats.dart
@@ -1,13 +1,4 @@
 class CacheStats {
-  CacheStats();
-
-  /// Memory bytes of cached item used, not precised
-  /// evicted item not traced.
-  int bytesInMemory = 0;
-
-  /// Count of miss in memory
-  int missInMemory = 0;
-
   /// Bytes of total local file
   int bytesInFile = 0;
 
@@ -25,12 +16,9 @@ class CacheStats {
 
   @override
   String toString() {
-    return "Bytes(Memory): $bytesInMemory\n"
-        "Miss(Memory): $missInMemory\n"
-        "Hit(Memory): $hitMemory\n"
-        "Hit(Files): $hitFiles\n"
-        "Bytes(File): $bytesInFile\n"
-        "Bytes Read from File: $bytesRead\n"
-        "Bytes download: $bytesDownload";
+    return 'Hit(Files): $hitFiles\n'
+        'Bytes(File): $bytesInFile\n'
+        'Bytes Read from File: $bytesRead\n'
+        'Bytes download: $bytesDownload';
   }
 }