+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.0
+
+* [break change] remove memory store
+* [break change] FileCache remove scan and async
+
 ## 0.4.0
 
 * Split to `file_cache` and `file_cache_flutter`